pyproject.toml
[project]
name = "animalclef26"
version = "0.1.0"
description = "AnimalCLEF 2026 - Wildlife Species Classification"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kaggle>=2.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "timm>=1.0.0",
    "wildlife-datasets>=1.0.0",
    "wildlife-tools @ git+https://github.com/WildlifeDatasets/wildlife-tools",
    "pillow>=9.0.0",
    "numpy>=1.23.0",
    "scikit-learn>=1.2.0",
    "umap-learn>=0.5.7",
    "pandas>=1.5.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "seaborn>=0.13.2",
]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
